requirements.txt
asgiref==3.7.2
Django==4.2.3
sqlparse==0.4.4
tzdata==2023.3
django-jazzmin==2.6.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
black==23.3.0
drf-spectacular==0.26.3
